+++ apps/time_master/src/db_inserter/model/time_sheet_model.hpp
@@ -27,11 +27,13 @@ struct DayData {
     int total_exercise_time;
     int cardio_time;
     int anaerobic_time;
-    // [删除] int exercise_both_time;
-    // [新增] 新增字段
     int gaming_time;
     int grooming_time;
     int toilet_time;
+    // --- [核心修改] 新增睡眠统计字段 ---
+    int sleep_night_time;
+    int sleep_day_time;
+    int sleep_total_time;
 };
 
 /**